Turn IncDecOp/SetOpOP into enum classes and add hhas support for them
I'm planning to make the opcode table support strongly-typed
OA args, so I'll convert the remaining OA's to have enum classes.
After this diff, only BareThis uses an OA arg without an associated
enum class (I'll change that if making OA take a type argument turns
out to be good).

Reviewed By: @swtaarrs

Differential Revision: D1097329

+++ hphp/runtime/vm/jit/ir.h
@@ -1000,7 +1000,7 @@ namespace constToBits_detail {
 
   template<class T>
   typename std::enable_if<needs_promotion<T>::value,uint64_t>::type
-  promoteIfNeeded(T t) { return t; }
+  promoteIfNeeded(T t) { return static_cast<uint64_t>(t); }
 
   template<class T>
   typename std::enable_if<!needs_promotion<T>::value,T>::type
